Shin-14 Add List of Active Users
Added a utils file containing utility functions. 
Added user join and user leave functions to populate a user array with a user objects of the form (username, socketid, roomId). The getRoomUsers function can then be called when a user joins a room or leaves a room, and the output parsed for usernames to populate the user list slide out.