SHIN-8 Handle End of Session
- Modified PeerJS and socket IO signals in chatjs and server js to handle video removal when a user leaves the room.
Known Bug: video does not remove properly when more than 2 users are in a room.

On join room only the user object is sent to server JS. The roomID, socketId and PeerId are all packaged into the user object.
Logs left in to show user_list in room on server side. and Peers and stream objects on client side

The Vid_div element on the page as been given an id of the stream Id to make it easily identifiable for removal.